Docs: Add explanation for exit code 137
This usually comes from the JVM rather than CodeQL itself.
On POSIX-compliant systems, this means the operating system sent signal 9, or `SIGKILL`, to terminate the CodeQL process.
This can happen when there is not enough available memory for CodeQL.
In this case, we can suggest a workaround to the user, or ask them to contact us if the problem persists.